requirements.txt
atproto
python-dotenv
pandas
pyarrow